autotest: Use c_uint64 to match platform for hash
Use uint64 c type when calculating hash to match the platform side
uint64_t. This explicit sizing makes hashing more reliable.

BUG=b:291991381
TEST=Ran autotest with a leased device that was flashed with CL with
platform side change.
`test_that --debug --no-retry --iterations 1 -b ${BOARD} 127.0.0.1:2200
autoupdate_MiniOS.with_exclusion.full  --args="skip_provisioning=True"`

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4718965
Tested-by: Saketh Pothireddy <spothire@google.com>
Commit-Queue: Saketh Pothireddy <spothire@google.com>

+++ client/cros/update_engine/update_engine_util.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from ctypes import c_size_t
+from ctypes import c_uint64
 import datetime
 import json
 import logging
@@ -880,9 +880,9 @@ class UpdateEngineUtil(object):
         @returns: The payload URL hash string as the exclusion name.
 
         """
-        result = c_size_t(0)
+        result = c_uint64(0)
         for c in payload_url:
-            result = c_size_t((result.value * 131) + ord(c))
+            result = c_uint64((result.value * 131) + ord(c))
         return str(result.value)
 
     def _apply_deferred_update(self, wait_for_idle=True):
